+++ bumblebee_py/intel/__init__.py
@@ -1,0 +1,1 @@
+"""Bumblebee threat-intel submodule for refreshing exposure catalogs."""